fix superhero file name
