--- src/Superheros.js
@@ -1,34 +1,0 @@
-const superheros = [
-    {
-      id: 1,
-      name: "ANT-MAN",
-      description: "Ability to shrink to sub-microscopic size"
-    },
-    {
-      id: 2,
-      name: "BATWOMAN",
-      description: "Highly skilled martial artis"
-    },
-    {
-      id: 3,
-      name: "CATWOMAN",
-      description: "Skilled gymnast"
-    },
-    {
-      id: 4,
-      name: "IRON MAN",
-      description: "Proficient scientist and engineer"
-    },
-    {
-      id: 5,
-      name: "BLACK PANTHER",
-      description: "Superhuman strength, speed, reflexes"
-    },
-    {
-      id: 6,
-      name: "CAPTAIN AMERICA",
-      description: "Agility, strength, speed, endurance"
-    }
-  ]
-
-export default superheros
